requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
colorama==0.4.6
confusable-homoglyphs==3.3.1
constantly==23.10.4
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
cryptography==46.0.3
daphne==4.2.1
Django==4.2.25
django-crispy-forms==2.4
django-formtools==2.5.1
django-otp==1.6.3
django-phonenumber-field==8.3.0
django-registration==5.2.1
django-two-factor-auth==1.18.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
dnspython==1.16.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
msgpack==1.1.2
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pymongo==3.11.4
pyOpenSSL==25.3.0
pytz==2025.2
qrcode==8.2
redis==7.0.1
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
zope.interface==8.0.1
python-dotenv==1.2.1